Changed typedef of H5::SizeType to match hsize_t exactly
* Added static_asserts in H5Test to check that HDF5 typedefs match

+++ src/complex/Utilities/Parsing/HDF5/H5.hpp
@@ -10,7 +10,7 @@ namespace H5
 {
 using IdType = int64_t;
 using ErrorType = int32_t;
-using SizeType = uint64_t;
+using SizeType = unsigned long long;
 
 inline const std::string DataTypeTag = "DataType";
 inline const std::string CompDims = "ComponentDimensions";
